backend/requirements.txt
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Validação de dados
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS e middleware
python-jose[cryptography]==3.3.0

# Banco de dados
supabase==2.0.3
python-dotenv==1.0.0

# Utilitários
python-dateutil==2.8.2